Pinning to Typescript 2.1.5
Typescript 2.2.0 is in RC and npm is alrady downloading it based on the previous dependency version rules.

VSCode is currently using 2.1.5, so it seems to make sense to follow that version, additinally the 2.2 compiler has a breaking change with lodash that prevents building spellchecker if tsc 2.2 is used.  cf.  https://github.com/DefinitelyTyped/DefinitelyTyped/pull/14502