Add rst to the list of enabled file types
I started using this spellchecker as I was writing documentation in
reStructuredText and I noticed rst wasn't listed as an enabled file type.
I dug around to see how to enable the spellchecker for rst documents
however it seemed as though the spellchecker is actually enabled for
reStructuredText out of the box.

I thought it might be helpful to include it in the list of enabled file
types in the readme.