feat: Add an autocorrect as you type option
This adds a new setting, `cSpell.autoCorrect`, that will automatically apply the first suggestion for a misspelled word as soon as the user types a non-word character (such as a space, comma, or semicolon).

The setting is off by default.

Disaclaimer: I don't know much about vscode extensions, so I used an LLM to help me with this PR.